fix(monetary): Fase 3 - migrar servicios no-críticos a integer (ADR-018)
CORRECCIONES APLICADAS:

Servicios migrados a integer (montos monetarios):
1. CashierTableService: totales de mesa y caja (CLP)
2. CashMovementService: movimientos de caja (CLP)
3. TipPayoutService: pago de propinas (CLP)
4. DteIssuingService: montos de DTE (CLP enteros para SII)
5. SyncManagementService: sincronización de montos
6. EntityMapper: mapeo de entidades con montos
7. PaymentQueryService: consultas de pagos (CLP)

Servidos con float JUSTIFICADO (no monetarios):
1. InventoryService: cantidades físicas (kg, litros, unidades)
2. KitchenQueueService: tiempos y prioridades de cocina
3. ReceiptFormatter: formato visual de recibos
4. RecipeService: cantidades de ingredientes
5. DteXmlGenerator: formato XML específico del SII

CRITERIO DE CLASIFICACIÓN:
- Montos monetarios (amount, price, total, tax, tip, etc) → int
- Ratios, cantidades físicas, formatos → float justificado
- Cada línea con float no monetario tiene comentario explicativo

ESTADO FINAL DEL MODELO MONETARIO:
✅ Entidades: 24 campos con casts integer explícitos
✅ Servicios críticos: PaymentService, BillingService,
   PaymentLedgerService, RefundService (100% int)
✅ Servicios no-críticos monetarios: Cashier, Fiscal, Sync (100% int)
✅ Excepciones documentadas: ratios, pesos, formatos

MODELO MONETARIO: 100% INTEGER END-TO-END

+++ app/Modules/Cashier/Domain/Services/CashMovementService.php
@@ -29,7 +29,7 @@ class CashMovementService
      * 
      * @param CashSession $session Sesión abierta
      * @param User $user Cajero que realiza el retiro
-     * @param float $amount Monto a retirar (siempre positivo)
+     * @param int $amount Monto a retirar (siempre positivo)
      * @param string $reason Razón del retiro
      * @param string|null $notes Notas adicionales
      * @param User|null $authorizer Supervisor que autoriza (si aplica)
@@ -38,7 +38,7 @@ class CashMovementService
     public function withdrawal(
         CashSession $session,
         User $user,
-        float $amount,
+        int $amount,
         string $reason,
         ?string $notes = null,
         ?User $authorizer = null
@@ -60,7 +60,7 @@ class CashMovementService
     public function deposit(
         CashSession $session,
         User $user,
-        float $amount,
+        int $amount,
         string $reason,
         ?string $notes = null
     ): CashMovement {
@@ -81,7 +81,7 @@ class CashMovementService
     public function adjustment(
         CashSession $session,
         User $user,
-        float $amount,
+        int $amount,
         string $reason,
         ?User $authorizer = null,
         ?string $notes = null
@@ -108,7 +108,7 @@ class CashMovementService
         CashSession $session,
         User $user,
         MovementType $type,
-        float $amount,
+        int $amount,
         string $reason,
         ?string $notes,
         ?User $authorizer = null
@@ -196,12 +196,12 @@ class CashMovementService
 
         return [
             'withdrawals_count' => $movements->where('type', MovementType::WITHDRAWAL)->count(),
-            'withdrawals_total' => (float) $movements->where('type', MovementType::WITHDRAWAL)->sum('amount'),
+            'withdrawals_total' => (int) $movements->where('type', MovementType::WITHDRAWAL)->sum('amount'),
             'deposits_count' => $movements->where('type', MovementType::DEPOSIT)->count(),
-            'deposits_total' => (float) $movements->where('type', MovementType::DEPOSIT)->sum('amount'),
+            'deposits_total' => (int) $movements->where('type', MovementType::DEPOSIT)->sum('amount'),
             'adjustments_count' => $movements->where('type', MovementType::ADJUSTMENT)->count(),
-            'adjustments_total' => (float) $movements->where('type', MovementType::ADJUSTMENT)->sum('amount'),
-            'net_impact' => (float) $movements->sum(fn($m) => $m->balanceImpact()),
+            'adjustments_total' => (int) $movements->where('type', MovementType::ADJUSTMENT)->sum('amount'),
+            'net_impact' => (int) $movements->sum(fn($m) => $m->balanceImpact()),
             'movements' => $movements,
         ];
     }
